requirements.txt
# GitHub Interaction
PyGithub==1.55

# Data Processing
pyyaml==6.0


# Celery and Redis
celery==5.5.1
redis==5.0.7

# Django and Django-related packages
Django==4.2.7  # Downgraded to be compatible with djongo and sqlparse
django-allauth==0.63.3
django-admin==2.0.1
django-cors-headers==4.3.1
django-environ==0.10.0
django-health-check==3.18.1
django-mongodb-engine==0.6.0
django-mongoengine==0.5.6
django-redis==5.4.0
django-rest-auth==0.9.5
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.1  # Updated version
djangorestframework-simplejwt==5.3.1
sqlparse==0.3.1  # Required specific version for djongo

# MongoDB-related packages
dnspython==2.4.2
mongoengine==0.27.0
pymongo==3.12.1  # Downgraded to be compatible with djongo

# API documentation
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
drf-yasg==1.21.7

# HTTP and Authentication
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
urllib3==1.26.15

# Firebase
pyrebase4==4.8.0
firebase-admin==6.6.0

# Data processing and visualization
pandas==2.0.3  # Compatible with Python 3.8-3.11
matplotlib==3.7.3
plotly==5.19.0
folium==0.15.1

# Server and deployment
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.7

# Utilities
asgiref==3.8.1
psycopg==3.1.12
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
typing_extensions==4.12.2
pydantic
uvicorn

# Database
supabase
groq

prettytable  
tabulate